first refactor, moved files around, changes some names
